strip lib dependency in the rlp module
