+++ modRlp/src/module-info.java
@@ -1,5 +1,3 @@
 module aion.rlp {
-    requires aion.util;
-
     exports org.aion.rlp;
 }
