+++ modRlp/src/org/aion/rlp/Hex.java
@@ -1,0 +1,104 @@
+package org.aion.rlp;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+
+/** Utility class for converting hex data to bytes and back again. */
+public class Hex {
+
+    private static final HexEncoder encoder = new HexEncoder();
+
+    public static String toHexString(byte[] data) {
+
+        return data == null ? "" : toHexString(data, 0, data.length);
+    }
+
+    public static String toHexString(byte[] data, int off, int length) {
+        byte[] encoded = encode(data, off, length);
+        return new String(encoded);
+    }
+
+    /**
+     * encode the input data producing a Hex encoded byte array.
+     *
+     * @return a byte array containing the Hex encoded data.
+     */
+    public static byte[] encode(byte[] data) {
+        return encode(data, 0, data.length);
+    }
+
+    /**
+     * encode the input data producing a Hex encoded byte array.
+     *
+     * @return a byte array containing the Hex encoded data.
+     */
+    public static byte[] encode(byte[] data, int off, int length) {
+        ByteArrayOutputStream bOut = new ByteArrayOutputStream();
+
+        encoder.encode(data, off, length, bOut);
+
+        return bOut.toByteArray();
+    }
+
+    /**
+     * Hex encode the byte data writing it to the given output stream.
+     *
+     * @return the number of bytes produced.
+     */
+    public static int encode(byte[] data, OutputStream out) {
+        return encoder.encode(data, 0, data.length, out);
+    }
+
+    /**
+     * Hex encode the byte data writing it to the given output stream.
+     *
+     * @return the number of bytes produced.
+     */
+    public static int encode(byte[] data, int off, int length, OutputStream out) {
+        return encoder.encode(data, off, length, out);
+    }
+
+    /**
+     * decode the Hex encoded input data. It is assumed the input data is valid.
+     *
+     * @return a byte array representing the decoded data.
+     */
+    public static byte[] decode(byte[] data) {
+        ByteArrayOutputStream bOut = new ByteArrayOutputStream();
+        try {
+            encoder.decode(data, 0, data.length, bOut);
+        } catch (IOException e) {
+            System.err.println("Hex decode failed! " + data);
+            return null;
+        }
+
+        return bOut.toByteArray();
+    }
+
+    /**
+     * decode the Hex encoded String data - whitespace will be ignored.
+     *
+     * @return a byte array representing the decoded data.
+     */
+    public static byte[] decode(String data) {
+        ByteArrayOutputStream bOut = new ByteArrayOutputStream();
+        try {
+            encoder.decode(data, bOut);
+        } catch (IOException e) {
+            System.err.println("Hex decode failed! " + data);
+            return null;
+        }
+        return bOut.toByteArray();
+    }
+
+    /**
+     * decode the Hex encoded String data writing it to the given output stream, whitespace
+     * characters will be ignored.
+     *
+     * @return the number of bytes produced.
+     */
+    public static int decode(String data, OutputStream out) throws IOException {
+        return encoder.decode(data, out);
+    }
+}
