+++ modRlp/src/org/aion/rlp/RLPList.java
@@ -1,7 +1,6 @@
 package org.aion.rlp;
 
 import java.util.ArrayList;
-import org.aion.util.bytes.ByteUtil;
 
 /**
  * @author Roman Mandeleil
@@ -11,9 +10,9 @@ public class RLPList extends ArrayList<RLPElement> implements RLPElement {
 
     private static final long serialVersionUID = -2855280911054117106L;
 
-    byte[] rlpData;
+    private byte[] rlpData;
 
-    public void setRLPData(byte[] rlpData) {
+    void setRLPData(byte[] rlpData) {
         this.rlpData = rlpData;
     }
 
@@ -21,7 +20,7 @@ public class RLPList extends ArrayList<RLPElement> implements RLPElement {
         return rlpData;
     }
 
-    public static void recursivePrint(RLPElement element) {
+    static void recursivePrint(RLPElement element) {
 
         if (element == null) {
             throw new RuntimeException("RLPElement object can't be null");
@@ -34,7 +33,7 @@ public class RLPList extends ArrayList<RLPElement> implements RLPElement {
             }
             System.out.print("]");
         } else {
-            String hex = ByteUtil.toHexString(element.getRLPData());
+            String hex = Hex.toHexString(element.getRLPData());
             System.out.print(hex + ", ");
         }
     }
