+++ modRlp/test/org/aion/rlp/ValueTest.java
@@ -8,7 +8,6 @@ import static org.junit.Assert.assertTrue;
 import java.math.BigInteger;
 import java.util.Arrays;
 import org.junit.Test;
-import org.spongycastle.util.encoders.Hex;
 
 public class ValueTest {
