+++ modUtil/build.gradle
@@ -3,4 +3,5 @@ ext.moduleName = 'aion.util'
 dependencies {
     testCompile 'junit:junit:4.12'
     testCompile 'pl.pragmatists:JUnitParams:1.1.1'
+    testCompile 'com.madgag.spongycastle:core:1.58.0.0'
 }
